endpoint: execute aliveCancel() before draining event queue
The context passed to ParseExternalRegenerationMetadata, which is the
one that ends up propagating all the way down to ReloadDatapath and
friends, can be Endpoint.aliveCtx if the ExternalRegenerationMetadata
does not contain a parent context.

Make sure this context is cancelled early on in the Endpoint teardown
process to allow downstream logic (bpf prog load, netlink attach, ..)
to bail out instead of trying to attach to resources that are already
removed by the container runtime.

Signed-off-by: Timo Beckers <timo@isovalent.com>