Fix an issue when 1st hop has real server but does not use VIP
In IPIP mode, we want to log VIP for external services. This is
fine when the 1st hop need to forward the packets to the node which
has real server. But in the 1st hop ( 1st node ) has the real server,
and no need to forward case, packets hit the pod using pod IP instead
of using VIP as dest IP. This patch fixed this by doing IPIP encap in
this case, and VIP will be used.

Signed-off-by: Jiang Wang <jiang.wang@bytedance.com>