add full port mapping support
Some services need to expose lots of ports (>500) to outside.
And it is tedious to add each port into the k8s config. To
address this issue, we add a full port mapping support. It
uses annotation to mark a full port range service, like the following:

 "metadata": {
        "annotations": {
            "io.cilium/full-port-mapping": "true"
        },

Once this annotation is set, cilium will not translate L4 ports for
external IPs and forward the packets directly to the backend pods.

One can verify this by running cilium service list cmd and check
for the outpus similar to the following:

70   192.168.122.238:0       FullPortsMapping   1 => 10.0.0.136:80
                                                2 => 10.0.2.175:80

Signed-off-by: Jiang Wang <jiang.wang@bytedance.com>