dsr: add a new config option ipipv4cni to separate ipv4 and ipv6 dsr
Previously, we (bytedance) reuse the old config -- "bpf-lb-dsr-dispatch: ipip"
to enable ipip dsr for ipv4. This one will cause ipv6 to not working
because that setting is applied for both ipv4 and ipv6, but ipv6
ipip is not changed to support cni mode (it can only be used in
lb-only mode). Also, we probably will not change ipv6 dsr to use ipip. As a
result, this commmit add a new value setting for ipip v4 dsr. That is

bpf-lb-dsr-dispatch: ipipv4cni

The new setting allow us to enable ipip dsr only for ipv4, and ipv6 dsr
will keep use the old default dsr mode, which is using ip options.

Signed-off-by: Jiang Wang <jiang.wang@bytedance.com>