ctmap/gc: do not terminate agent fatal on ctmap gc init after timeout
commit fea19ec9c730662c2c3f5ddd8f98092a3a5d140d upstream.

If the agent is not completed its initial ctmap/nat GC scan within a
hardcoded 30 second timeout then the agent with terminate with
level=fatal.

This can cause issues in situations where the agent is running on an
environment with high resource contention.  This makes it possible for
the initial gc to timeout and put Cilium into a crash-loop.

The only affects the ctmap pressure metrics controller, which isn't
critical so instead of terminating the agent will now wait for the
initial scan async, and will emit a warning log after 30 seconds if
initial gc scan has not completed.

This will allow us to still catch issues in CI/environments where it
is unlikely that we should see these timeouts, while still avoiding
unnecessary termination.

Fixes: #32680

Signed-off-by: Tom Hadlaw <tom.hadlaw@isovalent.com>

The context is different, make some code changes accordingly. We don't
have ct metric controller now.

Signed-off-by: Zijian Zhang <zijianzhang@bytedance.com>