pkg/endpoint: always copy existing state during synchronization
commit 900c8463f7c1e87db1c2e52ab481a9c61a5d220a upstream.

Endpoint regeneration goes to a lot of trouble to keep track of
whether the on-disk state has changed, only to avoid doing a couple
of readdir syscalls. The behaviour was added in commit  f6c4385a43
("pkg/endpoint: Keep BPF object files if compilation is skipped.")
and the message does not indicate that performance was of particular
concern.

Do the safe thing and always perform the state copy.

Signed-off-by: Lorenz Bauer <lmb@isovalent.com>

The context is different, make some code changes accordingly.

Signed-off-by: Zijian Zhang <zijianzhang@bytedance.com>